[WIP] huge refactoring
