+++ optimum/neuron/models/training/config.py
@@ -5,17 +5,12 @@ from typing import TYPE_CHECKING, Dict, Optional, Tuple, Union
 import torch
 
 from ...configuration_utils import NeuronConfig, register_neuron_config
-from ...distributed import ParallelizersManager
 from ...utils import is_neuronx_distributed_available
 from ...utils.torch_xla_and_neuronx_initialization import init_process_group
 
 
 if is_neuronx_distributed_available():
     from neuronx_distributed.parallel_layers import parallel_state
-
-
-if TYPE_CHECKING:
-    from transformers import PreTrainedModel
 
 
 @dataclass
@@ -73,23 +68,3 @@ class TrainingNeuronConfig(NeuronConfig):
     @property
     def should_parallelize(self):
         return self.tensor_parallel_size > 1 or self.pipeline_parallel_size > 1
-
-    def parallelize_model(
-        self,
-        model: "PreTrainedModel",
-        device: Optional["torch.device"] = None,
-    ) -> Union["PreTrainedModel", Tuple["PreTrainedModel", Dict[int, "torch.nn.Parameter"]]]:
-        parallelizer = ParallelizersManager.parallelizer_for_model(model)
-        parallelized_model = parallelizer.parallelize(
-            model,
-            device=device,
-            parallelize_embeddings=self.parallelize_embeddings,
-            sequence_parallel_enabled=self.sequence_parallel_enabled,
-            kv_size_multiplier=self.kv_size_multiplier,
-            pipeline_parallel_num_microbatches=self.pipeline_parallel_num_microbatches,
-            pipeline_parallel_use_zero1_optimizer=self.pipeline_parallel_use_zero1_optimizer,
-            pipeline_parallel_gradient_checkpointing_enabled=self.gradient_checkpointing,
-            checkpoint_dir=self.checkpoint_dir,
-            num_local_ranks_per_step=self.num_local_ranks_per_step,
-        )
-        return parallelized_model
