+++ optimum/neuron/models/training/pipeline_utils.py
@@ -24,7 +24,8 @@ from typing import Iterable
 import torch
 from torch import nn
 
-from ...distributed.utils import OptimumNeuronFXTracer
+from transformers.utils.fx import HFTracer
+
 from ...utils.import_utils import is_neuronx_distributed_available, is_torch_xla_available
 from .transformations_utils import get_tensor_model_parallel_attributes
 
@@ -41,6 +42,7 @@ if is_neuronx_distributed_available():
         get_pipeline_model_parallel_size,
     )
     from neuronx_distributed.pipeline import NxDPPModel
+    from neuronx_distributed.pipeline.trace import HFTracerWrapper, NxDTracer
 
 else:
 
@@ -50,6 +52,21 @@ else:
     class NxDPPModel:
         def __init__(self, *args, **kwargs):
             pass
+
+    class HFTracerWrapper:
+        def __init__(self, *args, **kwargs):
+            pass
+
+    class NxDTracer:
+        def __init__(self, *args, **kwargs):
+            pass
+
+
+class OptimumNeuronFXTracer(HFTracerWrapper):
+    def is_leaf_module(self, m: torch.nn.Module, module_qualified_name: str) -> bool:
+        return (
+            NxDTracer.is_leaf_module(self, m, module_qualified_name)
+            or HFTracer.is_leaf_module(self, m, module_qualified_name))
 
 
 class MetaParametersOnly:
